+++ devices/xml-filereader-old/xml-reader.cpp
@@ -50,7 +50,7 @@ struct timeval tv;
 
 	xml_Reader::xml_Reader (FILE		*f,
 	                        xmlDescriptor	*fd,
-	                        uint64_t	filePointer,
+	                        uint32_t	filePointer,
 	                        RingBuffer<std::complex<float>> *b,
 	                        bool	continue_on_eof) {
 	this	-> file		= f;
@@ -93,7 +93,7 @@ void	xml_Reader::stopReader	() {
 
 static	int cycleCount = 0;
 void	xml_Reader::run () {
-uint64_t	samplesRead	= 0;
+int	samplesRead	= 0;
 uint64_t	nextStop;
 int	startPoint	= filePointer;
 
@@ -137,9 +137,9 @@ int	startPoint	= filePointer;
 	}
 }
 
-uint64_t	xml_Reader::compute_nrSamples (FILE *f, int blockNumber) {
-uint64_t	nrElements	= fd -> blockList. at (blockNumber). nrElements;
-uint64_t	samplesToRead	= 0;
+int	xml_Reader::compute_nrSamples (FILE *f, int blockNumber) {
+int	nrElements	= fd -> blockList. at (blockNumber). nrElements;
+int	samplesToRead	= 0;
 
 	(void)f;
 	if (fd -> blockList. at (blockNumber). typeofUnit == "Channel") {
@@ -157,7 +157,7 @@ uint64_t	samplesToRead	= 0;
 	return samplesToRead;
 }
 
-uint64_t	xml_Reader::readSamples (FILE *theFile, 
+int	xml_Reader::readSamples (FILE *theFile, 
 	                         void(xml_Reader::*r)(FILE *theFile,
 	                                    std::complex<float> *, int)) {
 std::complex<float> temp [2048];
